Properly clean up stale commit task. Also reduce visibility of code.
