+++ searchlib/src/vespa/searchlib/transactionlog/domain.cpp
@@ -1,6 +1,8 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
 #include "domain.h"
+#include "domainpart.h"
+#include "session.h"
 #include <vespa/vespalib/util/stringfmt.h>
 #include <vespa/vespalib/util/closuretask.h>
 #include <vespa/vespalib/io/fileutil.h>
@@ -29,15 +31,6 @@ using std::make_shared;
 namespace search::transactionlog {
 
 VESPA_THREAD_STACK_TAG(domain_commit_executor);
-
-DomainConfig::DomainConfig()
-    : _encoding(Encoding::Crc::xxh64, Encoding::Compression::none),
-      _compressionLevel(9),
-      _fSyncOnCommit(false),
-      _partSizeLimit(0x10000000), // 256M
-      _chunkSizeLimit(0x40000),   // 256k
-      _chunkAgeLimit(10ms)
-{ }
 
 Domain::Domain(const string &domainName, const string & baseDir, Executor & commitExecutor,
                Executor & sessionExecutor, const DomainConfig & cfg, const FileHeaderContext &fileHeaderContext)
@@ -446,7 +439,7 @@ Domain::erase(SerialNum to)
 
 int
 Domain::visit(const Domain::SP & domain, SerialNum from, SerialNum to,
-              std::unique_ptr<Session::Destination> dest)
+              std::unique_ptr<Destination> dest)
 {
     assert(this == domain.get());
     cleanSessions();
