+++ searchlib/src/vespa/searchlib/transactionlog/domain.h
@@ -1,73 +1,23 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 #pragma once
 
-#include "domainpart.h"
-#include "session.h"
+#include "domainconfig.h"
+#include <vespa/vespalib/util/sync.h>
 #include <vespa/vespalib/util/threadexecutor.h>
-#include <vespa/vespalib/util/time.h>
-#include <vespa/fastos/thread.h>
-#include <chrono>
+#include <atomic>
 
+namespace search::common { class FileHeaderContext; }
 namespace search::transactionlog {
 
-class DomainConfig {
-public:
-    using duration = vespalib::duration;
-    DomainConfig();
-    DomainConfig & setEncoding(Encoding v)          { _encoding = v; return *this; }
-    DomainConfig & setPartSizeLimit(size_t v)       { _partSizeLimit = v; return *this; }
-    DomainConfig & setChunkSizeLimit(size_t v)      { _chunkSizeLimit = v; return *this; }
-    DomainConfig & setChunkAgeLimit(vespalib::duration v) { _chunkAgeLimit = v; return *this; }
-    DomainConfig & setCompressionLevel(uint8_t v)   { _compressionLevel = v; return *this; }
-    DomainConfig & setFSyncOnCommit(bool v)         { _fSyncOnCommit = v; return *this; }
-    Encoding          getEncoding() const { return _encoding; }
-    size_t       getPartSizeLimit() const { return _partSizeLimit; }
-    size_t      getChunkSizeLimit() const { return _chunkSizeLimit; }
-    duration     getChunkAgeLimit() const { return _chunkAgeLimit; }
-    uint8_t   getCompressionlevel() const { return _compressionLevel; }
-    bool         getFSyncOnCommit() const { return _fSyncOnCommit; }
-private:
-    Encoding     _encoding;
-    uint8_t      _compressionLevel;
-    bool         _fSyncOnCommit;
-    size_t       _partSizeLimit;
-    size_t       _chunkSizeLimit;
-    duration     _chunkAgeLimit;
-};
-
-struct PartInfo {
-    SerialNumRange range;
-    size_t numEntries;
-    size_t byteSize;
-    vespalib::string file;
-    PartInfo(SerialNumRange range_in, size_t numEntries_in, size_t byteSize_in, vespalib::stringref file_in)
-        : range(range_in),
-          numEntries(numEntries_in),
-          byteSize(byteSize_in),
-          file(file_in)
-    {}
-};
-
-struct DomainInfo {
-    using DurationSeconds = std::chrono::duration<double>;
-    SerialNumRange range;
-    size_t numEntries;
-    size_t byteSize;
-    DurationSeconds maxSessionRunTime;
-    std::vector<PartInfo> parts;
-    DomainInfo(SerialNumRange range_in, size_t numEntries_in, size_t byteSize_in, DurationSeconds maxSessionRunTime_in)
-        : range(range_in), numEntries(numEntries_in), byteSize(byteSize_in), maxSessionRunTime(maxSessionRunTime_in), parts() {}
-    DomainInfo()
-        : range(), numEntries(0), byteSize(0), maxSessionRunTime(), parts() {}
-};
-
-typedef std::map<vespalib::string, DomainInfo> DomainStats;
+class DomainPart;
+class Session;
 
 class Domain
 {
 public:
     using SP = std::shared_ptr<Domain>;
     using Executor = vespalib::SyncableThreadExecutor;
+    using DomainPartSP = std::shared_ptr<DomainPart>;
     Domain(const vespalib::string &name, const vespalib::string &baseDir, Executor & commitExecutor,
            Executor & sessionExecutor, const DomainConfig & cfg, const common::FileHeaderContext &fileHeaderContext);
 
@@ -78,7 +28,7 @@ public:
     bool erase(SerialNum to);
 
     void commit(const Packet & packet, Writer::DoneCallback onDone);
-    int visit(const Domain::SP & self, SerialNum from, SerialNum to, std::unique_ptr<Session::Destination> dest);
+    int visit(const Domain::SP & self, SerialNum from, SerialNum to, std::unique_ptr<Destination> dest);
 
     SerialNum begin() const;
     SerialNum end() const;
@@ -95,7 +45,7 @@ public:
     int closeSession(int sessionId);
 
     SerialNum findOldestActiveVisit() const;
-    DomainPart::SP findPart(SerialNum s);
+    DomainPartSP findPart(SerialNum s);
 
     static vespalib::string
     getDir(const vespalib::string & base, const vespalib::string & domain) {
@@ -138,8 +88,8 @@ private:
 
     SerialNumList scanDir();
 
-    using SessionList = std::map<int, Session::SP>;
-    using DomainPartList = std::map<int64_t, DomainPart::SP>;
+    using SessionList = std::map<int, std::shared_ptr<Session>>;
+    using DomainPartList = std::map<int64_t, DomainPartSP>;
     using DurationSeconds = std::chrono::duration<double>;
 
     DomainConfig           _config;
