+++ searchlib/src/vespa/searchlib/transactionlog/session.h
@@ -23,14 +23,6 @@ private:
     using time_point = std::chrono::time_point<std::chrono::steady_clock>;
 
 public:
-    class Destination {
-    public:
-        virtual ~Destination() = default;
-        virtual bool send(int32_t id, const vespalib::string & domain, const Packet & packet) = 0;
-        virtual bool sendDone(int32_t id, const vespalib::string & domain) = 0;
-        virtual bool connected() const = 0;
-        virtual bool ok() const = 0;
-    };
     typedef std::shared_ptr<Session> SP;
     Session(const Session &) = delete;
     Session & operator = (const Session &) = delete;
