+++ searchlib/src/vespa/searchlib/transactionlog/translogserverapp.h
@@ -23,11 +23,9 @@ private:
     void configure(std::unique_ptr<searchlib::TranslogserverConfig> cfg) override ;
 
 public:
-    typedef std::unique_ptr<TransLogServerApp> UP;
-
     TransLogServerApp(const config::ConfigUri & tlsConfigUri,
                       const common::FileHeaderContext &fileHeaderContext);
-    ~TransLogServerApp();
+    ~TransLogServerApp() override;
 
     TransLogServer::SP getTransLogServer() const;
